Update immich.yml (#13)
* Update immich.yml

Updating the compose file to be more in line with the current default for Immich. Added the saltbox managed label to each container. added user mapping to each container as well, as the majority of them will accept it. Removed default volumes in liu of the saltbox standard `/opt/appname` directories. Additional container added (typesense) since its also now included in the default.

* Update immich.yml

changed the `.env` file to reflect sandpit standards