feat(compose): Add Traefik configuration for immich server
- Added Traefik configuration for the immich server in the compose file
- Enabled Traefik and set up routing rules for different entrypoints
- Configured TLS options and certificate resolver for secure
communication
- Added load balancer server port for the immich service

This commit sets up Traefik configuration for the immich server in the
compose file. It enables Traefik and defines routing rules for the
immich service using different entrypoints. TLS options and a
certificate resolver are configured for secure communication.
Additionally, a load balancer server port is specified for the immich
service.

refactor(roles): Update default variables for immich server

- Updated default variables for the immich server role
- Added the necessary paths and folders for the immich server containers
- Set up web-related variables for the immich web container
- Configured DNS records and proxy settings for domain resolution
- Adjusted Traefik settings and enabled or disabled the API endpoint

This commit updates the default variables for the immich server role. It
includes the necessary paths and folders for the immich server to
function properly. Web-related variables for the immich web container
are set up, including the subdomain, domain, and port. DNS records and
proxy settings are configured for domain resolution. Additionally,
Traefik settings are adjusted, such as enabling or disabling the API
endpoint.

chore(compose/readme): Add notes for using the immich server role in the
sandbox

- Added notes for using the immich server role in the sandbox
- Provided instructions for setting up user mapping and creating
subdirectories
- Included additional steps for configuring the name change and running
the application

This commit adds notes to the compose/readme/Immich.md file for using
the immich server role in the sandbox. It provides instructions for
setting up user mapping and creating the necessary subdirectories.
Additional steps are included for configuring the name change from admin
to another chosen name and running the immich application.

chore(roles/immich): Create default variables for immich server role

- Created default variables file for the immich server role
- Defined variables for paths, web configuration, DNS settings, Traefik,
and Docker
- Set initial values for these variables to be customized as needed

This commit creates a default variables file for the immich server role.
It defines variables for various aspects of the server, including paths,
web configuration, DNS settings, Traefik configuration, and Docker
settings. Initial values are set for these variables, which can be
customized as needed.