Fix Immich documentation and remove unnecessary containers (#5086)
The Immich documentation has been updated to reflect the changes in
version 0.88.*. The immich web and immich proxy containers are no longer
used, and their functionality has been moved to the Immich server.
Instructions have been added to create an A record in Cloudflare or use
the ddns role to automatically create the records. The compose file's
`labels` section should be updated with the domain and top-level domain.
Other user-specific configurations should also be reviewed and adjusted
as necessary.